requirements.txt
fastapi==0.116.1
yfinance==0.2.65
pandas==2.3.2
apscheduler==3.11.0
python-dotenv==1.1.1
uvicorn==0.35.0
numpy==2.3.2
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
PyYAML==6.0.2